faulthy aliases in view definition
the aliases in the subquery for the view vaccess where v and w, in different collations this is a problem as v and w can be seen as the same character and the sql parser fail the script. I changed the alias v to v1 and the alias w to w2.